+++ src/Home/Drawer/Drawer.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode, useContext } from "react";
 import { View, StyleSheet, Text } from "react-native";
 import LinearGradient from "react-native-linear-gradient";
 
-import { logout } from "../Authentication";
+import { logout } from "../../Authentication";
 import {
   RoundedIcon,
   ChartIcon,
@@ -10,14 +10,14 @@ import {
   FriendsIcon,
   ProfileIcon,
   SearchIcon,
-} from "../Components";
-import { BrokenLogout } from "../Components/CustomIcons";
+} from "../../Components";
+import { BrokenLogout } from "../../Components/CustomIcons";
 import {
   DrawerNavigationParamList,
   DrawerProps,
-} from "../Components/Navigation";
-import { FontType, Size, wWidth } from "../Components/StyleConstants";
-import { UserConext } from "../Context";
+} from "../../Components/Navigation";
+import { FontType, Size, wWidth } from "../../Components/StyleConstants";
+import { UserConext } from "../../Context";
 
 import Item from "./Item";
 import User from "./User";
