+++ src/Home/Drawer/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { createDrawerNavigator } from "@react-navigation/drawer";
 import { StyleSheet } from "react-native";
 
-import { DrawerNavigationParamList } from "../Components/Navigation";
 import { Search } from "../Search";
+import { DrawerNavigationParamList } from "../../Components/Navigation";
 
 import Drawer, { DRAWER_WIDTH } from "./Drawer";
